Player collectableDiamonds attribute
