+++ Workspace/DiamondKill/src/com/github/calfur/minecraftserverplugins/diamondkill/database/PlayerData.java
@@ -1,4 +1,4 @@
-package com.github.calfur.minecraftserverplugins.diamondkill;
+package com.github.calfur.minecraftserverplugins.diamondkill.database;
 
 import java.util.HashMap;
 import java.util.Map;
